Add withdraw/deiconify and reformat
Yes, that was autopep formatting. I used `black .` and got this. Personally, I don't like the autopep style. Should we just remove that?

The messagebox dialog isn't centered for some reason anymore though.

The self.resize_app() on line 226 does nothing, because it calls the main windows resize, not the settings window. Maybe a special center/resize function in a misc file is in order?