feat, refactor: settings window, code cleanup(?) (#21)
* some updates
WHY DO WE USE PYOWM? DATA RECEIVED FROM REQUESTS DOES THE JOB WELL, AND EVERYTHING IS IN ONE LIST/DICTIONARY!

* feat, refactor: icons for buttons, code arrangement

* feat: settings window

* update: window now always on top and focused

* reset some things

* refactor: reducing lines for string statements, removing image icons

* refactor: change `OWMCITY` to `owm_search`
It's no longer a single function running the app.

* fix: typos in `README.md`

* Add withdraw/deiconify and reformat

Yes, that was autopep formatting. I used `black .` and got this. Personally, I don't like the autopep style. Should we just remove that?

The messagebox dialog isn't centered for some reason anymore though.

The self.resize_app() on line 226 does nothing, because it calls the main windows resize, not the settings window. Maybe a special center/resize function in a misc file is in order?

---------

Co-authored-by: Moosems <95927277+Moosems@users.noreply.github.com>